backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
numpy==1.26.4
pandas==2.2.3
python-dotenv==1.0.1
google-genai>=1.0.0
httpx>=0.27.0
